Fixed bug in Function.uber() construction, better testClass.html
Originally in the subclass constructor, by the time the uber function
was called, var 'prop' would hold its last value from the loop, and
so every uber() function would call the same property.
To fix it, used a closure to hold the value of prop at every iteration
of the for-in loop.

Also changed testClass.html to demonstrate the constructor and uber()
capabilities of Class.js